app/pipeline/pyproject.toml
[project]
name = "pipeline"
requires-python = ">=3.10,<3.15"
version = "0.1.0"
dependencies = [
    "dagster==1.13.0",
    "duckdb",
    "google-genai",
    "pydantic",
    "pandas",
    "tabulate",
    "tqdm>=4.67.3",
    "questionary>=2.1.1",
]

[dependency-groups]
dev = [
    "dagster-webserver",
    "dagster-dg-cli",
    "pytest",
    "pytest-asyncio",
    "pytest-mock",
    "ruff",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
force-include = { "pyproject.toml" = "pyproject.toml" }

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "pipeline"
registry_modules = [
    "pipeline.components.*",
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.lint.per-file-ignores]
"src/pipeline/prompts/*" = ["E501"]  # LLM prompt strings are naturally long
"src/pipeline/utils/judge.py" = ["E501", "E402"]  # Long prompt strings + deferred imports
"scripts/*" = ["E402"]  # sys.path manipulation required before imports

[tool.ruff.format]
quote-style = "double"

